airplan_game/requirements.txt
pygame>=2.1.0
Pillow>=9.0.0
pyinstaller>=6.0.0
